partitioned the Filter's Modal body section into a component
