+++ client-side/src/components/Navbar/Navbar.css
@@ -311,12 +311,20 @@ font-size: 12px;
   margin-top: 13px;
  }
  .clearAll {
-  margin-right: 335px;
   font-weight: 600;
   text-decoration: underline;
+  margin-left: -150px;
  }
  .colorBorder {
-  background-color: rgb(242, 227, 227);
+  background-color: rgb(227, 208, 208);
   border:  2px solid black;
+ }
+ .footerButton {
+  width: 162px;
+  height: 48px;
+  background-color: black;
+  color: white;
+  font-weight: bold;
+  border-radius: 15px;
  }
